Размещение проекта по папкам и вынисение логики из фрагмента поля в ViewModel. ViewModel реализуется посредствам интерфейса, для возможности назначения разных ViewModel длф фрагмента, в зависимости от состояния игры.
