+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.PlayGame">
         <activity
-            android:name=".PlacementRoomActivity"
+            android:name=".Activity.PlacementRoomActivity"
             android:screenOrientation="landscape">
             <!--            <intent-filter>-->
             <!--                <action android:name="android.intent.action.MAIN" />-->
@@ -18,11 +18,11 @@
             <!--                <category android:name="android.intent.category.LAUNCHER" />-->
             <!--            </intent-filter>-->
         </activity>
-        <activity android:name=".UserPageActivity" />
+        <activity android:name=".Activity.UserPageActivity" />
         <activity
-            android:name=".RoomActivity"
+            android:name=".Activity.RoomActivity"
             android:screenOrientation="landscape" />
-        <activity android:name=".MainActivity">
+        <activity android:name=".Activity.MainActivity">
 
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -31,8 +31,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".SignUpActivity" />
-        <activity android:name=".DashboardActivity" />
+        <activity android:name=".Activity.SignUpActivity" />
+        <activity android:name=".Activity.DashboardActivity" />
     </application>
 
 </manifest>
