+++ app/src/main/java/by/bsuir/playgame/Activity/SignUpActivity.java
@@ -1,4 +1,4 @@
-package by.bsuir.playgame;
+package by.bsuir.playgame.Activity;
 
 import android.app.ProgressDialog;
 import android.content.Intent;
@@ -17,6 +17,8 @@ import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
+
+import by.bsuir.playgame.R;
 
 public class SignUpActivity extends AppCompatActivity {
     private EditText emailEt,passwordEt1,passwordEt2;
