+++ app/src/main/java/by/bsuir/playgame/Activity/UserPageActivity.java
@@ -1,10 +1,9 @@
-package by.bsuir.playgame;
+package by.bsuir.playgame.Activity;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
-import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
@@ -18,6 +17,8 @@ import com.google.firebase.database.Query;
 import com.google.firebase.database.ValueEventListener;
 
 import java.util.Objects;
+
+import by.bsuir.playgame.R;
 
 public class UserPageActivity extends AppCompatActivity {
