+++ app/src/main/java/by/bsuir/playgame/Interfece/IFieldViewModel.java
@@ -1,0 +1,9 @@
+package by.bsuir.playgame.Interfece;
+
+import androidx.lifecycle.LiveData;
+
+public interface IFieldViewModel {
+    void setPoint(String point);
+    LiveData<int[]> getIcon();
+
+}
