+++ app/src/main/res/layout-land/activity_room.xml
@@ -10,7 +10,7 @@
 
     <fragment
         android:id="@+id/fragment1"
-        android:name="by.bsuir.playgame.FieldFragment"
+        android:name="by.bsuir.playgame.Fragment.FieldFragment"
         android:layout_width="300dp"
         android:layout_height="300dp"
         android:layout_alignParentStart="true"
@@ -29,7 +29,7 @@
 
     <fragment
         android:id="@+id/fragment2"
-        android:name="by.bsuir.playgame.FieldFragment"
+        android:name="by.bsuir.playgame.Fragment.FieldFragment"
         android:layout_width="300dp"
         android:layout_height="300dp"
         android:layout_alignParentStart="true"
@@ -37,13 +37,4 @@
         android:layout_marginStart="80dp"
         android:layout_marginTop="48dp" />
 
-    <Button
-        android:id="@+id/button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="514dp"
-        android:layout_marginTop="89dp"
-        android:text="Button" />
 </RelativeLayout>
