+++ app/src/main/res/layout/activity_main.xml
@@ -8,7 +8,7 @@
     android:orientation="vertical"
     android:paddingHorizontal="30dp"
     android:background="#1E88E5"
-    tools:context=".MainActivity">
+    tools:context=".Activity.MainActivity">
 
     <TextView
         android:layout_width="wrap_content"
