+++ app/src/main/res/layout/activity_room.xml
@@ -5,7 +5,7 @@
     android:id="@+id/layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RoomActivity">
+    tools:context=".Activity.RoomActivity">
 
     <Button
         android:id="@+id/click"
