+++ app/src/main/res/layout/activity_user_page.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".UserPageActivity" >
+    tools:context=".Activity.UserPageActivity" >
 
     <EditText
         android:id="@+id/NameUser"
